clobbers app ports with integration ports (#3)
* clobbers app ports with integration ports

When upping services that expose ports to the host, there is a high
potential for port conflicts. In order to eliminate port conflicts, it
is necessary to clobber the ports that the application's docker-compose
file describes with a range of ports that the nib-integrate gem
provides. These ports start with the `initial_port` key found in the gem
configuration file. The default is 10_000.

Additionally, it is necessary to create a docker-compose file that is
essentially empty in the directory of the application. The only contents
of the "empty" file is the `version` key. The "empty" config file gets
the version of docker-compose from the application's docker-compose.yml
file. The reason this is necessary is that docker-compose uses the
location of the docker-compose.yml file to determine local paths (such
as environment files).

In order for the port clobbering to work, a complete copy of the app's
docker-compose.yml file is created, and new ports are assigned for
services that have a "ports" key.

The command now removes the .nib-integrate* files (empty config and
config copy) from the application directory.