presentation for docker network and nib-integrate
I was asked to give a lunch and learn talk on docker networking and how
nib integrate works with it to integrate multiple docker-compose
applications into a single network in the development environment.

The index.html in the presentation directory can be opened in a browser
and used as a presentation. The presentation is written in reveal.js